+++ CHANGELOG.rst
@@ -29,6 +29,9 @@ Fixed
 - Fix errors in edge cases (mainly when `n_total` equals `ref_total`)
   in the computation of difference of two proportions confidence intervals
   using `wang` method.
+- Fix errors in generating latex table for risk report (function ``make_risk_report``):
+  - Escape special characters in latex.
+  - Fix the header for the generated latex table.
 
 Security
 ~~~~~~~~